// src/lib/json/types/Construction.ts
export interface IConstruction {
    hq: number
    bz: number
    fz: number
    sp: number
    fe: number
    kr: number
    fr: number
    or: number
    fo: number
    go: number
    fel: number
    krl: number
    frl: number
    orl: number
    fol: number
    gol: number
    kkw: number
    fkw: number
    rsf: number
    vst: number
    spa: number
    fwa: number
    ha: number
    ah: number
    hp: number
    hz: number
    ba: number
    ki: number
    gdz: number
    ws: number
}